Fix IResourceName aliasing
And alter readme
