+++ src/main/java/com/google/api/codegen/transformer/PathTemplateTransformer.java
@@ -14,17 +14,12 @@
  */
 package com.google.api.codegen.transformer;
 
-import com.google.api.codegen.config.FieldConfig;
-import com.google.api.codegen.config.FieldModel;
 import com.google.api.codegen.config.FixedResourceNameConfig;
 import com.google.api.codegen.config.InterfaceConfig;
 import com.google.api.codegen.config.MethodConfig;
 import com.google.api.codegen.config.ResourceNameConfig;
-import com.google.api.codegen.config.ResourceNameMessageConfigs;
 import com.google.api.codegen.config.ResourceNameOneofConfig;
-import com.google.api.codegen.config.ResourceNameType;
 import com.google.api.codegen.config.SingleResourceNameConfig;
-import com.google.api.codegen.util.Name;
 import com.google.api.codegen.viewmodel.FormatResourceFunctionView;
 import com.google.api.codegen.viewmodel.ParseResourceFunctionView;
 import com.google.api.codegen.viewmodel.PathTemplateGetterFunctionView;
@@ -35,18 +30,10 @@ import com.google.api.codegen.viewmodel.ResourceNameOneofView;
 import com.google.api.codegen.viewmodel.ResourceNameParamView;
 import com.google.api.codegen.viewmodel.ResourceNameSingleView;
 import com.google.api.codegen.viewmodel.ResourceNameView;
-import com.google.api.codegen.viewmodel.ResourceProtoFieldView;
-import com.google.api.codegen.viewmodel.ResourceProtoView;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ListMultimap;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Set;
 
 /** PathTemplateTransformer generates view objects for path templates from a service model. */
@@ -180,64 +167,6 @@ public class PathTemplateTransformer {
             .index(index)
             .value(config.getFixedValue());
     return builder.build();
-  }
-
-  public List<ResourceProtoView> generateResourceProtos(GapicInterfaceContext context) {
-    SurfaceNamer namer = context.getNamer();
-    ResourceNameMessageConfigs resourceConfigs =
-        context.getProductConfig().getResourceNameMessageConfigs();
-    ListMultimap<String, FieldModel> fieldsByMessage =
-        resourceConfigs.getFieldsWithResourceNamesByMessage();
-    Map<String, FieldConfig> fieldConfigMap =
-        context.getProductConfig().getDefaultResourceNameFieldConfigMap();
-    List<ResourceProtoView> protos = new ArrayList<>();
-    for (Entry<String, Collection<FieldModel>> entry : fieldsByMessage.asMap().entrySet()) {
-      String msgName = entry.getKey();
-      Collection<FieldModel> fields = new ArrayList<>(entry.getValue());
-      ResourceProtoView.Builder protoBuilder = ResourceProtoView.newBuilder();
-      protoBuilder.protoClassName(namer.getTypeNameConverter().getTypeName(msgName).getNickname());
-      List<ResourceProtoFieldView> fieldViews = new ArrayList<>();
-      for (FieldModel field : fields) {
-        FieldConfig fieldConfig = fieldConfigMap.get(field.getFullName());
-        String fieldTypeSimpleName = namer.getResourceTypeName(fieldConfig.getResourceNameConfig());
-        String fieldTypeName =
-            context
-                .getImportTypeTable()
-                .getAndSaveNicknameForTypedResourceName(fieldConfig, fieldTypeSimpleName);
-        if (field.isRepeated()) {
-          fieldTypeName = fieldTypeName.replaceFirst("scg::IEnumerable", "gax::ResourceNameList");
-        }
-        String fieldDocTypeName = fieldTypeName.replace('<', '{').replace('>', '}');
-        String fieldElementTypeName =
-            context
-                .getImportTypeTable()
-                .getAndSaveNicknameForResourceNameElementType(fieldConfig, fieldTypeSimpleName);
-        ResourceProtoFieldView fieldView =
-            ResourceProtoFieldView.newBuilder()
-                .typeName(fieldTypeName)
-                .parseMethodTypeName(namer.getPackageName() + "." + fieldTypeName)
-                .docTypeName(fieldDocTypeName)
-                .elementTypeName(fieldElementTypeName)
-                .isAny(fieldConfig.getResourceNameType() == ResourceNameType.ANY)
-                .isRepeated(field.isRepeated())
-                .isOneof(fieldConfig.getResourceNameType() == ResourceNameType.ONEOF)
-                .propertyName(namer.getResourceNameFieldGetFunctionName(fieldConfig))
-                .underlyingPropertyName(namer.publicMethodName(Name.from(field.getSimpleName())))
-                .build();
-        fieldViews.add(fieldView);
-      }
-      protoBuilder.fields(fieldViews);
-      protos.add(protoBuilder.build());
-    }
-    Collections.sort(
-        protos,
-        new Comparator<ResourceProtoView>() {
-          @Override
-          public int compare(ResourceProtoView a, ResourceProtoView b) {
-            return a.protoClassName().compareTo(b.protoClassName());
-          }
-        });
-    return protos;
   }
 
   public List<FormatResourceFunctionView> generateFormatResourceFunctions(
