+++ src/test/java/com/google/api/codegen/testdata/csharp/csharp_library.baseline
@@ -12774,9 +12774,9 @@ namespace Google.Example.Library.V1
     public partial class GetBookFromAbsolutelyAnywhereRequest
     {
         /// <summary>
-        /// <see cref="IResourceName"/>-typed view over the <see cref="Name"/> resource name property.
+        /// <see cref="gax::IResourceName"/>-typed view over the <see cref="Name"/> resource name property.
         /// </summary>
-        public IResourceName AsResourceName
+        public gax::IResourceName AsResourceName
         {
             get { return string.IsNullOrEmpty(Name) ? null : gax::UnknownResourceName.Parse(Name); }
             set { Name = value != null ? value.ToString() : ""; }
@@ -12861,9 +12861,9 @@ namespace Google.Example.Library.V1
     public partial class ListStringsRequest
     {
         /// <summary>
-        /// <see cref="IResourceName"/>-typed view over the <see cref="Name"/> resource name property.
+        /// <see cref="gax::IResourceName"/>-typed view over the <see cref="Name"/> resource name property.
         /// </summary>
-        public IResourceName AsResourceName
+        public gax::IResourceName AsResourceName
         {
             get { return string.IsNullOrEmpty(Name) ? null : gax::UnknownResourceName.Parse(Name); }
             set { Name = value != null ? value.ToString() : ""; }
@@ -12874,10 +12874,10 @@ namespace Google.Example.Library.V1
     public partial class ListStringsResponse
     {
         /// <summary>
-        /// <see cref="gax::ResourceNameList{IResourceName}"/>-typed view over the <see cref="Strings"/> resource name property.
+        /// <see cref="gax::ResourceNameList{gax::IResourceName}"/>-typed view over the <see cref="Strings"/> resource name property.
         /// </summary>
-        public gax::ResourceNameList<IResourceName> StringsAsResourceNames =>
-            new gax::ResourceNameList<IResourceName>(Strings,
+        public gax::ResourceNameList<gax::IResourceName> StringsAsResourceNames =>
+            new gax::ResourceNameList<gax::IResourceName>(Strings,
                 str => gax::UnknownResourceName.Parse(str));
 
     }
